chatbot/requirements.txt
Flask==3.1.0
openai==1.66.2
requests==2.32.3
python-dotenv==1.0.1
